+++ packages/core/package.json
@@ -13,7 +13,9 @@
     "./admin-dashboard.ui": "./src/admin-dashboard.ui.ts",
     "./member-profile.server": "./src/member-profile.server.ts",
     "./member-profile.ui": "./src/member-profile.ui.ts",
-    "./admin.server": "./src/modules/admin/index.server.ts",
+    "./admins": "./src/modules/admin/admin.core.ts",
+    "./admins.types": "./src/modules/admin/admin.types.ts",
+    "./admins.ui": "./src/modules/admin/admin.ui.tsx",
     "./applications": "./src/modules/application/application.core.ts",
     "./applications.ui": "./src/modules/application/application.ui.tsx",
     "./education": "./src/modules/education/education.core.ts",
