+++ packages/core/src/admin-dashboard.ui.ts
@@ -1,5 +1,4 @@
 export { BullQueue } from './infrastructure/bull/bull.types';
-export { AddAdminInput } from './modules/admin/admin.types';
 export {
   SendOneTimeCodeInput,
   VerifyOneTimeCodeInput,
