+++ packages/core/src/modules/admin/admin.types.ts
@@ -1,24 +1,14 @@
 import { z } from 'zod';
 
-import { Email, Entity } from '@oyster/types';
+import { BooleanInput, Email } from '@oyster/types';
 
-// Schemas
+// Use Cases
 
-export const Admin = Entity.extend({
+export const AddAdminInput = z.object({
   email: Email,
   firstName: z.string().trim().min(1),
-  isAmbassador: z.boolean().default(false),
+  isAmbassador: BooleanInput.default(false),
   lastName: z.string().trim().min(1),
 });
 
-export const AddAdminInput = Admin.pick({
-  email: true,
-  firstName: true,
-  lastName: true,
-  isAmbassador: true,
-});
-
-// Types
-
-export type Admin = z.infer<typeof Admin>;
 export type AddAdminInput = z.infer<typeof AddAdminInput>;
