--- packages/core/src/modules/admin/index.server.ts
@@ -1,1 +1,0 @@
-export { isMemberAdmin } from './queries/is-member-admin';
