--- packages/core/src/modules/admin/queries/is-member-admin.ts
@@ -1,10 +1,0 @@
-import { db } from '@oyster/db';
-
-export async function isMemberAdmin(memberId: string) {
-  const admin = await db
-    .selectFrom('admins')
-    .where('memberId', '=', memberId)
-    .executeTakeFirst();
-
-  return !!admin;
-}
